后端bug修复

+++ back-client/src/main/java/org/example/backclient/entity/Response.java
@@ -2,7 +2,7 @@ package org.example.backclient.entity;
 
 import com.alibaba.fastjson2.JSONObject;
 
-public record Response(int id, int code, Object data, String message) {
+public record Response(long id, int code, Object data, String message) {
     public boolean success() { return code == 200;}
     public JSONObject toJSON() { return JSONObject.from(data);}
     public String asString() {return data.toString();}
